+++ lib/src/shared.dart
@@ -1,9 +1,11 @@
+import 'package:break_infinity/src/helpers.dart';
+
 /// For example: if two exponents are more than 17 apart,
 /// consider adding them together pointless, just return the larger one
 const int maxSignificantDigits = 17;
 
 /// Highest value you can safely put here is `0x7FFFFFFFFFFFFFFF - MAX_SIGNIFICANT_DIGITS` where the int.max is represented as the
-const int expLimit = intMaxValue - maxSignificantDigits;
+final int expLimit = maxInt - maxSignificantDigits;
 
 /// The largest exponent that can appear in a Number, though not all mantissas are valid here.
 const int numberExpMax = 308;
@@ -12,10 +14,12 @@ const int numberExpMax = 308;
 const int numberExpMin = -324;
 
 /// Tolerance which is used for Number conversion to compensate floating-point error.
-const double roundTolerance = 1e-10;
+final double roundTolerance = isJavaScript ? 1e-10 : 1e-18;
 
-/// Represents the smallest 64 bit integer on native platforms supported by Dart
-const int intMinValue = -0x8000000000000000;
+// The following is adapted from https://github.com/dart-lang/sdk/issues/41717#issuecomment-622312466
 
-/// Represents the largest 64 bit integer on native platforms
-const int intMaxValue = 0x7FFFFFFFFFFFFFFF;
+/// Represents the largest integer on platforms supported by Dart
+final int maxInt = isJavaScript ? double.infinity as int : ~minInt;
+
+/// Represents the smallest integer on platforms supported by Dart
+final int minInt = isJavaScript ? -double.infinity as int : (-1 << 63);